move x_bogux_.py, absolute imports in uploader.py
