+++ TikTokUploader/uploader.py
@@ -2,10 +2,10 @@ import requests
 import json
 import time
 import datetime
-from x_bogus_ import get_x_bogus
 from urllib.parse import urlencode
 
-from .util import assertSuccess, printError, getTagsExtra, uploadToTikTok, log, getCreationId
+from TikTokUploader.x_bogus_ import get_x_bogus
+from TikTokUploader.util import assertSuccess, printError, getTagsExtra, uploadToTikTok, log, getCreationId
 
 
 UA = 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/68.0.3440.106 Safari/537.36'
